refactor: moving interface to proper namespace
